+++ lib/src/stream_sink_transformer.dart
@@ -53,6 +53,8 @@ abstract class StreamSinkTransformer<S, T> {
   /// This means that calls to [StreamSink.add] on the returned sink may throw a
   /// [CastError] if the argument type doesn't match the reified type of the
   /// sink.
+  @deprecated
+  // TODO remove TypeSafeStreamSinkTransformer
   static StreamSinkTransformer<S, T> typed<S, T>(
           StreamSinkTransformer transformer) =>
       transformer is StreamSinkTransformer<S, T>